made particle Type with variable precision
