+++ src/autopas/particles/Particle.h
@@ -22,9 +22,10 @@ namespace autopas {
  * This class can be used to build your own Particle class. However, you are
  * free to not use it as well.
  */
-class Particle {
+template <typename floatType>
+class ParticleBase {
  public:
-  Particle() : _r({0.0, 0.0, 0.0}), _v({0., 0., 0.}), _f({0.0, 0.0, 0.0}), _id(0) {}
+  ParticleBase() : _r({0.0, 0.0, 0.0}), _v({0., 0., 0.}), _f({0.0, 0.0, 0.0}), _id(0) {}
 
   /**
    * Constructor of the Particle class
@@ -32,34 +33,34 @@ class Particle {
    * @param v velocity of the particle
    * @param id id of the particle
    */
-  Particle(std::array<double, 3> r, std::array<double, 3> v, unsigned long id)
+  ParticleBase(std::array<floatType, 3> r, std::array<floatType, 3> v, unsigned long id)
       : _r(r), _v(v), _f({0.0, 0.0, 0.0}), _id(id) {}
 
-  virtual ~Particle() = default;
+  virtual ~ParticleBase() = default;
 
   /**
    * get the force acting on the particle
    * @return force
    */
-  const std::array<double, 3> &getF() const { return _f; }
+  const std::array<floatType, 3> &getF() const { return _f; }
 
   /**
    * Set the force acting on the particle
    * @param f force
    */
-  void setF(const std::array<double, 3> &f) { _f = f; }
+  void setF(const std::array<floatType, 3> &f) { _f = f; }
 
   /**
    * Add a partial force to the force acting on the particle
    * @param f partial force to be added
    */
-  void addF(const std::array<double, 3> &f) { _f = ArrayMath::add(_f, f); }
+  void addF(const std::array<floatType, 3> &f) { _f = ArrayMath::add(_f, f); }
 
   /**
    * Substract a partial force from the force acting on the particle
    * @param f partial force to be substracted
    */
-  void subF(const std::array<double, 3> &f) { _f = ArrayMath::sub(_f, f); }
+  void subF(const std::array<floatType, 3> &f) { _f = ArrayMath::sub(_f, f); }
 
   /**
    * Get the id of the particle
@@ -77,37 +78,37 @@ class Particle {
    * Get the position of the particle
    * @return current position
    */
-  const std::array<double, 3> &getR() const { return _r; }
+  const std::array<floatType, 3> &getR() const { return _r; }
 
   /**
    * Set the position of the particle
    * @param r new position
    */
-  void setR(const std::array<double, 3> &r) { _r = r; }
+  void setR(const std::array<floatType, 3> &r) { _r = r; }
 
   /**
    * Add a distance vector to the position of the particle
    * @param r vector to be added
    */
-  void addR(const std::array<double, 3> &r) { _r = ArrayMath::add(_r, r); }
+  void addR(const std::array<floatType, 3> &r) { _r = ArrayMath::add(_r, r); }
 
   /**
    * Get the velocity of the particle
    * @return current velocity
    */
-  const std::array<double, 3> &getV() const { return _v; }
+  const std::array<floatType, 3> &getV() const { return _v; }
 
   /**
    * Set the velocity of the particle
    * @param v new velocity
    */
-  void setV(const std::array<double, 3> &v) { _v = v; }
+  void setV(const std::array<floatType, 3> &v) { _v = v; }
 
   /**
    * Add a vector to the current velocity of the particle
    * @param v vector to be added
    */
-  void addV(const std::array<double, 3> &v) { _v = ArrayMath::add(_v, v); }
+  void addV(const std::array<floatType, 3> &v) { _v = ArrayMath::add(_v, v); }
 
   /**
    * Creates a string containing all data of the particle.
@@ -136,31 +137,45 @@ class Particle {
   /**
    * the type for the soa storage
    */
-  typedef autopas::utils::SoAType<size_t, double, double, double, double, double, double>::Type SoAArraysType;
+  typedef
+      typename autopas::utils::SoAType<size_t, floatType, floatType, floatType, floatType, floatType, floatType>::Type
+          SoAArraysType;
 
   /**
    * The Storage Arrays for Cuda
    */
-  typedef autopas::utils::CudaSoAType<size_t, double, double, double, double, double, double>::Type
-      CudaDeviceArraysType;
+  typedef typename autopas::utils::CudaSoAType<size_t, floatType, floatType, floatType, floatType, floatType,
+                                               floatType>::Type CudaDeviceArraysType;
 
  protected:
   /**
    * Particle position as 3D coordinates.
    */
-  std::array<double, 3> _r;
+  std::array<floatType, 3> _r;
   /**
    * Particle velocity as 3D vector.
    */
-  std::array<double, 3> _v;
+  std::array<floatType, 3> _v;
   /**
    * Force the particle experiences as 3D vector.
    */
-  std::array<double, 3> _f;
+  std::array<floatType, 3> _f;
   /**
    * Particle id.
    */
   unsigned long _id;
 };
+/**
+ * Particle with all variables in 32 bit precision
+ */
+typedef ParticleBase<float> ParticleFP32;
+/**
+ * Particle with all variables in 64 bit precision
+ */
+typedef ParticleBase<double> ParticleFP64;
+/**
+ * Alias for Particle with all variables in 64 bit precision
+ */
+typedef ParticleFP64 Particle;
 
 }  // namespace autopas
